修改了分区上面分类的点击事件，暂时将里面的详情页面的轮播图去除

+++ app/src/main/java/com/bilibili/lingxiao/play/model/VideoData.kt
@@ -29,12 +29,12 @@ data class VideoData(
         val x_80: List<X80>
     ) {
         data class X32(
-            val filesize: Int,
+            val filesize: Long,
             val timelength: Int
         )
 
         data class X16(
-            val filesize: Int,
+            val filesize: Long,
             val timelength: Int
         )
 
@@ -44,12 +44,12 @@ data class VideoData(
         )
 
         data class X64(
-            val filesize: Int,
+            val filesize: Long,
             val timelength: Int
         )
 
         data class X80(
-            val filesize: Int,
+            val filesize: Long,
             val timelength: Int
         )
     }
